+++ src/Fauna/create_herbivore_individual.h
@@ -17,7 +17,7 @@ class CreateHerbivoreIndividual : public CreateHerbivoreCommon {
  public:
   /// Constructor
   /** \copydoc CreateHerbivoreCommon::CreateHerbivoreCommon() */
-  CreateHerbivoreIndividual(const Hft* hft,
+  CreateHerbivoreIndividual(const std::shared_ptr<const Hft> hft,
                             const std::shared_ptr<const Parameters> parameters)
       : CreateHerbivoreCommon(hft, parameters) {}