+++ src/Fauna/world.cpp
@@ -54,9 +54,8 @@ void World::create_simulation_unit(std::shared_ptr<Habitat> habitat) {
   PopulationList* populations = new PopulationList();
 
   // Fill the object with one population per HFT.
-  for (int i = 0; i < get_hfts().size(); i++) {
-    const Hft* phft = &get_hfts()[i];
-    populations->add(world_constructor->create_population(phft));
+  for (const auto& hft : get_hfts()) {
+    populations->add(world_constructor->create_population(hft));
   }
   assert(populations != NULL);
 
@@ -124,11 +123,11 @@ void World::simulate_day(const Date& date, const bool do_herbivores) {
     if (establish_if_needed)
       for (const auto& hft : get_hfts()) {
         PopulationList& pops = sim_unit.get_populations();
-        if (!pops.exists(hft))
-          pops.add(world_constructor->create_population(&hft));
+        if (!pops.exists(*hft))
+          pops.add(world_constructor->create_population(hft));
 
-        PopulationInterface& p = pops.get(hft);
-        if (p.get_list().empty()) pops.get(hft).establish();
+        PopulationInterface& p = pops.get(*hft);
+        if (p.get_list().empty()) pops.get(*hft).establish();
       }
 
     // Create function object to delegate all simulations for this day to.